Move idalib_ea2str() from strings_extras.h to kernwin_extras.h (but keep it in the `strings` module for the time being)
